% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_names.R
\docType{data}
\name{element_closeness_str}
\alias{element_closeness_str}
\title{UI display string for element_closeness variable}
\format{
An object of class \code{character} of length 1.
}
\usage{
element_closeness_str
}
\description{
UI display string for element_closeness variable
}
\keyword{datasets}
