% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_names.R
\docType{data}
\name{element_cluster_ID_str}
\alias{element_cluster_ID_str}
\title{UI display string for element_cluster variable}
\format{
An object of class \code{character} of length 1.
}
\usage{
element_cluster_ID_str
}
\description{
UI display string for element_cluster variable
}
\keyword{datasets}
