% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_names.R
\docType{data}
\name{element_network_degree_norm_str}
\alias{element_network_degree_norm_str}
\title{UI display string for element_ndn variable}
\format{
An object of class \code{character} of length 1.
}
\usage{
element_network_degree_norm_str
}
\description{
UI display string for element_ndn variable
}
\keyword{datasets}
