% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_names.R
\docType{data}
\name{mineral_closeness_str}
\alias{mineral_closeness_str}
\title{UI display string for mineral_closeness variable}
\format{
An object of class \code{character} of length 1.
}
\usage{
mineral_closeness_str
}
\description{
UI display string for mineral_closeness variable
}
\keyword{datasets}
