% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_names.R
\docType{data}
\name{mineral_cluster_ID_str}
\alias{mineral_cluster_ID_str}
\title{UI display string for mineral_cluster variable}
\format{
An object of class \code{character} of length 1.
}
\usage{
mineral_cluster_ID_str
}
\description{
UI display string for mineral_cluster variable
}
\keyword{datasets}
