% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_dragon}
\alias{run_dragon}
\title{Run the "dragon" Shiny Application. Wrapper for dragon::run_app().}
\usage{
run_dragon()
}
\description{
Run the "dragon" Shiny Application. Wrapper for dragon::run_app().
}
\examples{
\dontrun{
library(dragon)
dragon::run_dragon()
}

}
