% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_timeline.R
\docType{data}
\name{selected_time_frame_levels}
\alias{selected_time_frame_levels}
\title{Timeplot plot labels for user colors}
\format{
An object of class \code{character} of length 2.
}
\usage{
selected_time_frame_levels
}
\description{
Timeplot plot labels for user colors
}
\keyword{datasets}
