% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare-med-data.R
\name{try_url}
\alias{try_url}
\title{Function to check for internet and read a specified html}
\usage{
try_url(url)
}
\arguments{
\item{url}{The URL to read}
}
\value{
Website read with `xml2::read_html`, or FALSE if could not reach internet/website
}
\description{
Function to check for internet and read a specified html
}
